Finishes off scaffolding the SqlServer project and initial aspects of the testing projects. This commit is to finish off the "concept" aspect of the feature so that we can merge into "develop" and  start a more formal feature branching based on actual trackable work items.

+++ Fabric.Terminology.Domain/Models/IPagerSettings.cs
@@ -2,7 +2,7 @@
 {
     public interface IPagerSettings
     {
-        long CurrentPage { get; set; }
-        long ItemsPerPage { get; set; }        
+        int CurrentPage { get; set; }
+        int ItemsPerPage { get; set; }        
     }
 }
